Privacy update to json and final lists conversion to work with a11y
Restruted Privacy.tsx to support the move to JSON. Removed old privacy files as this has now been converted over to json to work with the new a11y-list functionality for screen readers > Can now be found in the "en" and "fr" json files within the "local" directory. Update the remaining lists to work correctly with the new a11y-list functionality.