Fixes blank tutorial pages when in landscape
This fixes an issue where tutorial pages would render as blank when presented in landscape mode.

The `Carousel` component would clip subviews that would be rendered offscreen and when the view would be presented, it would not be re-rendered. As the entirety of the view is offscreen when presented in landscape mode, it was not rendered until the user interacted with the view.