privacy policy a11y intergration
Updating a11y list functionality to work with the privacy policy. Still some work left to do, but the privacy policy is having to be converted over to json and work like the rest of the content in the app. This is being done due to the fact it is not possible to correct the list accessibility issues with how it is being generated visa the markup format