Updated FR content for future usage statistics
Same as PR #1387 for French file: Updated content for coming usage statistics on the menu, province or territory selection screens and steps for sharing exposures if positive. Also removed the previous content about new notifications, but I'm not sure if something else needs to be done to delete the white card? Please review the figma page carefully to ensure I didn't miss anything.