App transport security fix (#1449)
* Update fastlane to 2.176.0 (was 2.174.0) - also updates
aws-partitions 1.429.0 (was 1.427.0)
aws-sdk-s3 1.89.0 (was 1.88.1)
emoji_regex 3.2.2 (was 3.2.1)

* Add a new fastlane lane that checks to see that there is no NSAppTransportSecurity key in the final IPA
This check is only performed when a :build_and_deploy with production. This action can be triggered by the yarn deploy-ios:production action. @cds-snc team need to confirm if this is sufficient for checking the build that will end up in the appstore

* Remove the NSAppTransportSecurity key & dict from the Info.plist